+++ Modules/Ruleset.Firewall/Public/Find-RulePrincipal.ps1
@@ -31,13 +31,27 @@ SOFTWARE.
 Get all firewall rules with or without LocalUser value
 
 .DESCRIPTION
-Get all rules which are either missing or not missing LocalUser value
-Rules which are missing LocalUser are considered weak and need to be updated
+Get all rules which are either missing or not missing LocalUser value, and save the result
+into a JSON file.
+Rules which are missing LocalUser are considered weak and need to be updated.
 This operation is slow, intended for debugging.
 
-.PARAMETER Empty
-If specified, returns rules with no local user value
+.PARAMETER Path
+Path into which to save file.
+Wildcard characters are supported.
+
+.PARAMETER FileName
+Output file name, which is json file into which result is saved
+
+.PARAMETER Append
+Append exported rules to existing file instead of replacing
+
+.PARAMETER Weak
+If specified, returns rules with no local user value,
 Otherwise only rules with local user are returned
+
+.PARAMETER Direction
+Firewall rule direction, default is '*' both directions
 
 .EXAMPLE
 PS> Find-RulePrincipal -Empty
@@ -46,37 +60,128 @@ PS> Find-RulePrincipal -Empty
 None. You cannot pipe objects to Find-RulePrincipal
 
 .OUTPUTS
-TODO: Describe outputs and OutputType
+[System.Void]
 
 .NOTES
-TODO: This needs improvement to export matching rules to JSON
+None.
 #>
 function Find-RulePrincipal
 {
 	[CmdletBinding(
 		HelpURI = "https://github.com/metablaster/WindowsFirewallRuleset/blob/master/Modules/Ruleset.Firewall/Help/en-US/Find-RulePrincipal.md")]
+	[OutputType([void])]
 	param (
+		[Parameter(Mandatory = $true)]
+		[SupportsWildcards()]
+		[System.IO.DirectoryInfo] $Path,
+
 		[Parameter()]
-		[switch] $Empty
+		[string] $FileName = "NoPrincipalRules",
+
+		[Parameter()]
+		[switch] $Append,
+
+		[Parameter()]
+		[switch] $Weak,
+
+		[Parameter()]
+		[ValidateSet("Inbound", "Outbound", "*")]
+		[string] $Direction = "*"
 	)
 
 	Write-Debug -Message "[$($MyInvocation.InvocationName)] ParameterSet = $($PSCmdlet.ParameterSetName):$($PSBoundParameters | Out-String)"
 
+	$StopWatch = [System.Diagnostics.Stopwatch]::new()
+	$StopWatch.Start()
+
+	# Exclude rules for store apps
 	Write-Information -Tags $MyInvocation.InvocationName -MessageData "INFO: Getting rules from GPO..."
-	$GPORules = Get-NetFirewallProfile -All -PolicyStore $PolicyStore |
-	Get-NetFirewallRule | Where-Object -Property Owner -EQ $null
+	$GPORules = Get-NetFirewallRule -PolicyStore $PolicyStore |
+	Where-Object {
+		$null -eq $_.Owner -and
+		$_.Direction -like $Direction
+	}
 
+	# Exclude rules with LocalUser set\unset
 	Write-Information -Tags $MyInvocation.InvocationName -MessageData "INFO: Applying security filter..."
-	$UserFilter = $GPORules | Get-NetFirewallSecurityFilter |
-	Where-Object -Property LocalUser -EQ Any
 
+	if ($Weak)
+	{
+		$UserFilter = $GPORules | Get-NetFirewallSecurityFilter |
+		Where-Object -Property LocalUser -EQ Any
+	}
+	else
+	{
+		$UserFilter = $GPORules | Get-NetFirewallSecurityFilter |
+		Where-Object -Property LocalUser -NE Any
+	}
+
+	# Exclude rules for services because these can't have LocalUser property set
 	Write-Information -Tags $MyInvocation.InvocationName -MessageData "INFO: Applying service filter..."
 	$ServiceFilter = $UserFilter | Get-NetFirewallRule |
 	Get-NetFirewallServiceFilter | Where-Object -Property Service -EQ Any
 
 	Write-Information -Tags $MyInvocation.InvocationName -MessageData "INFO: Selecting properties..."
-	$TargetRules = $ServiceFilter | Get-NetFirewallRule |
-	Select-Object -Property DisplayName, DisplayGroup, Direction
+	$ResultRules = $ServiceFilter | Get-NetFirewallRule |
+	Select-Object -Property DisplayName, DisplayGroup, Direction |
+	Sort-Object -Property DisplayGroup
 
-	$TargetRules | Format-List
+	$StopWatch.Stop()
+	$TotalHours = $StopWatch.Elapsed | Select-Object -ExpandProperty Hours
+	$TotalMinutes = $StopWatch.Elapsed | Select-Object -ExpandProperty Minutes
+	$TotalMinutes += $TotalHours * 60
+
+	Write-Information -Tags $MyInvocation.InvocationName -MessageData "INFO: Time needed to find weak rules was: $TotalMinutes minutes"
+
+	# Replace 1 and 2 with Inbound and Outbound
+	$TargetRules = @()
+	foreach ($Rule in $ResultRules)
+	{
+		$TargetRules += [PSCustomObject]@{
+			DisplayName = $Rule.DisplayName
+			DisplayGroup = $Rule.DisplayGroup
+			Direction = if ($Rule.Direction -eq 1) { "Inbound" } else { "Outbound" }
+		}
+	}
+
+	$Path = Resolve-FileSystemPath $Path -Create
+	if (!$Path)
+	{
+		# Errors if any, reported by Resolve-FileSystemPath
+		return
+	}
+
+	# NOTE: Split-Path -Extension is not available in Windows PowerShell
+	$FileExtension = [System.IO.Path]::GetExtension($FileName)
+
+	# Save result to JSON file
+	if (!$FileExtension -or ($FileExtension -ne ".json"))
+	{
+		Write-Debug -Message "[$($MyInvocation.InvocationName)] Adding extension to input file"
+		$FileName += ".json"
+	}
+
+	Write-Information -Tags $MyInvocation.InvocationName -MessageData "INFO: Exporting result to $Path\$FileName"
+
+	if ($Append)
+	{
+		if (Test-Path -PathType Leaf -Path "$Path\$FileName")
+		{
+			Write-Debug -Message "[$($MyInvocation.InvocationName)] Appending result to JSON file"
+			$JsonFile = ConvertFrom-Json -InputObject (Get-Content -Path "$Path\$FileName" -Raw)
+
+			@($JsonFile; $TargetRules) | ConvertTo-Json |
+			Set-Content -Path "$Path\$FileName" -Encoding $DefaultEncoding
+		}
+		else
+		{
+			Write-Warning -Message "Not appending result to file because no existing file"
+			$TargetRules | ConvertTo-Json | Set-Content -Path "$Path\$FileName" -Encoding $DefaultEncoding
+		}
+	}
+	else
+	{
+		Write-Debug -Message "[$($MyInvocation.InvocationName)] Replacing content in JSON file"
+		$TargetRules | ConvertTo-Json | Set-Content -Path "$Path\$FileName" -Encoding $DefaultEncoding
+	}
 }