+++ Templates/Manifest.Module/Public/New-Function.ps1
@@ -27,7 +27,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 #>
 
-# TODO: Remove using statement
+# TODO: Remove or update using statement
 using namespace System
 
 <#